Adds support for recursive parent tweet lookups
This is useful if someone, for example, asks for the sauce of an image in a tweet, and another user replies to that comment by mentioning saucenao.

Currently, the bot would just try and look up the sauce of the comment asking for the sauce, which won't work obviously! Now we will recursively look up the comment chain to find the closest parent tweet with media and look up the sauce to that.