Logging added to aid in future development.
Ideally, we want to abandon using trace.moe's similarity as a reference entirely and just make sure the titles match. But in order to do this, we will need to perform some basic parsing on titles to account for differences in the way things like season numbers and release years are displayed between databases.